+++ src/lib.rs
@@ -190,6 +190,7 @@ The minimum supported version of Rust for `glam` is `1.45.0`.
 
 #[macro_use]
 mod macros;
+
 #[macro_use]
 mod vec;
 
@@ -200,6 +201,7 @@ mod affine2;
 mod affine3;
 mod core;
 mod euler;
+mod features;
 mod mat2;
 mod mat3;
 mod mat4;
@@ -209,7 +211,6 @@ mod vec3;
 mod vec4;
 mod vec_mask;
 
-mod features;
 #[cfg(feature = "spirv-std")]
 mod spirv;
 
@@ -228,7 +229,7 @@ pub use self::bool::*;
 /** `f32` vector, quaternion and matrix types. */
 pub mod f32 {
     pub use super::affine2::Affine2;
-    pub use super::affine3::Affine3;
+    pub use super::affine3::Affine3A;
     pub use super::mat2::{mat2, Mat2};
     pub use super::mat3::{mat3, mat3a, Mat3, Mat3A};
     pub use super::mat4::{mat4, Mat4};
@@ -238,7 +239,8 @@ pub mod f32 {
     pub use super::vec4::{vec4, Vec4};
 
     #[cfg(feature = "transform-types")]
-    pub use super::transform::{TransformRt, TransformSrt};
+    #[allow(deprecated)]
+    pub use super::transform::{TransformRT, TransformSRT};
 }
 pub use self::f32::*;