+++ src/transform.rs
@@ -1,4 +1,6 @@
-use crate::{Mat4, Quat, Vec3, Vec3A, Vec3Swizzles, Vec4Swizzles};
+#![allow(deprecated)]
+
+use crate::{Affine3A, Mat4, Quat, Vec3, Vec3A, Vec3Swizzles};
 use core::ops::Mul;
 
 #[cfg(feature = "rand")]
@@ -7,43 +9,54 @@ use rand::{
     Rng,
 };
 
-#[derive(Clone, Copy, PartialEq, PartialOrd, Debug)]
+/**
+ * A transform containing non-uniform scale, rotation and translation.
+ *
+ * Scale and translation are stored as `Vec3A` for better performance.
+ */
+#[derive(Clone, Copy, PartialEq, Debug)]
 #[repr(C)]
-pub struct TransformSrt {
+#[deprecated(
+    since = "0.15.0",
+    note = "Moving to a separate crate, see https://github.com/bitshifter/glam-rs/issues/175"
+)]
+pub struct TransformSRT {
+    pub rotation: Quat,
+    pub translation: Vec3,
     pub scale: Vec3,
-    pub rotation: Quat,
-    pub translation: Vec3,
 }
 
-impl Default for TransformSrt {
+impl Default for TransformSRT {
     #[inline]
     fn default() -> Self {
-        Self {
-            scale: Vec3::ONE,
-            rotation: Quat::IDENTITY,
-            translation: Vec3::ZERO,
-        }
+        Self::IDENTITY
     }
 }
 
-#[derive(Clone, Copy, PartialEq, PartialOrd, Debug)]
+/**
+ * A transform containing rotation and translation.
+ *
+ * Translation is stored as a `Vec3A` for better performance.
+ */
+#[derive(Clone, Copy, PartialEq, Debug)]
 #[repr(C)]
-pub struct TransformRt {
+#[deprecated(
+    since = "0.15.0",
+    note = "Moving to a separate crate, see https://github.com/bitshifter/glam-rs/issues/175"
+)]
+pub struct TransformRT {
     pub rotation: Quat,
     pub translation: Vec3,
 }
 
-impl Default for TransformRt {
+impl Default for TransformRT {
     #[inline]
     fn default() -> Self {
-        Self {
-            rotation: Quat::IDENTITY,
-            translation: Vec3::ZERO,
-        }
+        Self::IDENTITY
     }
 }
 
-impl TransformSrt {
+impl TransformSRT {
     /// The identity transforms that does nothing.
     pub const IDENTITY: Self = Self {
         scale: Vec3::ONE,
@@ -54,18 +67,9 @@ impl TransformSrt {
     #[inline]
     pub fn from_scale_rotation_translation(scale: Vec3, rotation: Quat, translation: Vec3) -> Self {
         Self {
-            scale,
             rotation,
             translation,
-        }
-    }
-
-    #[inline]
-    pub fn from_transform_rt(scale: Vec3, rt: &TransformRt) -> Self {
-        Self {
             scale,
-            rotation: rt.rotation,
-            translation: rt.translation,
         }
     }
 
@@ -75,9 +79,9 @@ impl TransformSrt {
         let rotation = self.rotation.conjugate();
         let translation = -(rotation * (self.translation * scale));
         Self {
-            scale,
             rotation,
             translation,
+            scale,
         }
     }
 
@@ -107,12 +111,22 @@ impl TransformSrt {
 
     #[inline]
     pub fn transform_point3(&self, other: Vec3) -> Vec3 {
-        (self.rotation * (other * self.scale)) + self.translation
+        self.transform_point3a(other.into()).into()
     }
 
     #[inline]
     pub fn transform_vector3(&self, other: Vec3) -> Vec3 {
-        self.rotation * (other * self.scale)
+        self.transform_vector3a(other.into()).into()
+    }
+
+    #[inline]
+    pub fn transform_point3a(&self, other: Vec3A) -> Vec3A {
+        (self.rotation * (other * Vec3A::from(self.scale))) + Vec3A::from(self.translation)
+    }
+
+    #[inline]
+    pub fn transform_vector3a(&self, other: Vec3A) -> Vec3A {
+        self.rotation * (other * Vec3A::from(self.scale))
     }
 
     /// Returns true if the absolute difference of all elements between `self`
@@ -135,7 +149,7 @@ impl TransformSrt {
 }
 
 #[inline]
-fn mul_srt_srt(lhs: &TransformSrt, rhs: &TransformSrt) -> TransformSrt {
+fn mul_srt_srt(lhs: &TransformSRT, rhs: &TransformSRT) -> TransformSRT {
     // Based on https://github.com/nfrechette/rtm `rtm::qvv_mul`
     let lhs_scale = Vec3A::from(lhs.scale);
     let rhs_scale = Vec3A::from(rhs.scale);
@@ -145,50 +159,46 @@ fn mul_srt_srt(lhs: &TransformSrt, rhs: &TransformSrt) -> TransformSrt {
     if min_scale.cmplt(Vec3A::ZERO).any() {
         // If negative scale, we go through a matrix
         let lhs_mtx =
-            Mat4::from_scale_rotation_translation(lhs.scale, lhs.rotation, lhs.translation);
+            Affine3A::from_scale_rotation_translation(lhs.scale, lhs.rotation, lhs.translation);
         let rhs_mtx =
-            Mat4::from_scale_rotation_translation(rhs.scale, rhs.rotation, rhs.translation);
+            Affine3A::from_scale_rotation_translation(rhs.scale, rhs.rotation, rhs.translation);
         let mut result_mtx = lhs_mtx * rhs_mtx;
 
         let sign = scale.signum();
-        result_mtx.x_axis = (Vec3A::from(result_mtx.x_axis).normalize() * sign.xxx()).extend(0.0);
-        result_mtx.y_axis = (Vec3A::from(result_mtx.y_axis).normalize() * sign.yyy()).extend(0.0);
-        result_mtx.z_axis = (Vec3A::from(result_mtx.z_axis).normalize() * sign.zzz()).extend(0.0);
+        result_mtx.x_axis = result_mtx.x_axis.normalize() * sign.xxx();
+        result_mtx.y_axis = result_mtx.y_axis.normalize() * sign.yyy();
+        result_mtx.z_axis = result_mtx.z_axis.normalize() * sign.zzz();
 
         let scale = Vec3::from(scale);
-        let rotation = Quat::from_rotation_mat4(&result_mtx);
-        let translation = result_mtx.w_axis.xyz();
-        TransformSrt {
-            scale,
+        let rotation = Quat::from_affine3(&result_mtx);
+        let translation = Vec3::from(result_mtx.translation);
+        TransformSRT {
             rotation,
             translation,
+            scale,
         }
     } else {
-        let scale = Vec3::from(scale);
         let rotation = lhs.rotation * rhs.rotation;
-        let translation = Vec3::from(
-            (rhs.rotation * (Vec3A::from(lhs.translation) * rhs_scale))
-                + Vec3A::from(rhs.translation),
-        );
-        TransformSrt {
-            scale,
+        let translation = (rhs.rotation * (lhs.translation * rhs.scale)) + rhs.translation;
+        TransformSRT {
             rotation,
             translation,
+            scale: scale.into(),
         }
     }
 }
 
 #[inline]
-fn mul_rt_rt(lhs: &TransformRt, rhs: &TransformRt) -> TransformRt {
+fn mul_rt_rt(lhs: &TransformRT, rhs: &TransformRT) -> TransformRT {
     let rotation = lhs.rotation * rhs.rotation;
     let translation = (rhs.rotation * lhs.translation) + rhs.translation;
-    TransformRt {
+    TransformRT {
         rotation,
         translation,
     }
 }
 
-impl TransformRt {
+impl TransformRT {
     /// The identity transforms that does nothing.
     pub const IDENTITY: Self = Self {
         rotation: Quat::IDENTITY,
@@ -250,11 +260,21 @@ impl TransformRt {
 
     #[inline]
     pub fn transform_point3(&self, other: Vec3) -> Vec3 {
-        (self.rotation * other) + self.translation
+        self.transform_point3a(other.into()).into()
     }
 
     #[inline]
     pub fn transform_vector3(&self, other: Vec3) -> Vec3 {
+        self.transform_vector3a(other.into()).into()
+    }
+
+    #[inline]
+    pub fn transform_point3a(&self, other: Vec3A) -> Vec3A {
+        (self.rotation * other) + Vec3A::from(self.translation)
+    }
+
+    #[inline]
+    pub fn transform_vector3a(&self, other: Vec3A) -> Vec3A {
         self.rotation * other
     }
 
@@ -276,31 +296,15 @@ impl TransformRt {
     }
 }
 
-impl Mul<Vec3> for TransformRt {
-    type Output = Vec3;
+impl Mul<TransformRT> for TransformRT {
+    type Output = TransformRT;
     #[inline]
-    fn mul(self, other: Vec3) -> Vec3 {
-        self.transform_point3(other)
-    }
-}
-
-impl Mul<Vec3> for TransformSrt {
-    type Output = Vec3;
-    #[inline]
-    fn mul(self, other: Vec3) -> Vec3 {
-        self.transform_point3(other)
-    }
-}
-
-impl Mul<TransformRt> for TransformRt {
-    type Output = TransformRt;
-    #[inline]
-    fn mul(self, other: TransformRt) -> TransformRt {
+    fn mul(self, other: TransformRT) -> TransformRT {
         mul_rt_rt(&self, &other)
     }
 }
 
-impl Mul<TransformSrt> for TransformSrt {
+impl Mul<TransformSRT> for TransformSRT {
     type Output = Self;
     #[inline]
     fn mul(self, other: Self) -> Self::Output {
@@ -308,25 +312,25 @@ impl Mul<TransformSrt> for TransformSrt {
     }
 }
 
-impl Mul<TransformRt> for TransformSrt {
-    type Output = TransformSrt;
+impl Mul<TransformRT> for TransformSRT {
+    type Output = TransformSRT;
     #[inline]
-    fn mul(self, other: TransformRt) -> Self::Output {
+    fn mul(self, other: TransformRT) -> Self::Output {
         mul_srt_srt(&self, &other.into())
     }
 }
 
-impl Mul<TransformSrt> for TransformRt {
-    type Output = TransformSrt;
+impl Mul<TransformSRT> for TransformRT {
+    type Output = TransformSRT;
     #[inline]
-    fn mul(self, other: TransformSrt) -> Self::Output {
+    fn mul(self, other: TransformSRT) -> Self::Output {
         mul_srt_srt(&self.into(), &other)
     }
 }
 
-impl From<TransformRt> for TransformSrt {
+impl From<TransformRT> for TransformSRT {
     #[inline]
-    fn from(tr: TransformRt) -> Self {
+    fn from(tr: TransformRT) -> Self {
         Self {
             translation: tr.translation,
             rotation: tr.rotation,
@@ -336,10 +340,10 @@ impl From<TransformRt> for TransformSrt {
 }
 
 #[cfg(feature = "rand")]
-impl Distribution<TransformRt> for Standard {
+impl Distribution<TransformRT> for Standard {
     #[inline]
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> TransformRt {
-        TransformRt::from_rotation_translation(
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> TransformRT {
+        TransformRT::from_rotation_translation(
             rng.gen::<Quat>(),
             Vec3::new(
                 rng.gen_range(core::f32::MIN..=core::f32::MAX),
@@ -351,16 +355,16 @@ impl Distribution<TransformRt> for Standard {
 }
 
 #[cfg(feature = "rand")]
-impl Distribution<TransformSrt> for Standard {
+impl Distribution<TransformSRT> for Standard {
     #[inline]
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> TransformSrt {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> TransformSRT {
         let mut gen_non_zero = || loop {
             let f: f32 = rng.gen_range(core::f32::MIN..=core::f32::MAX);
             if f.abs() > core::f32::MIN_POSITIVE {
                 return f;
             }
         };
-        TransformSrt::from_scale_rotation_translation(
+        TransformSRT::from_scale_rotation_translation(
             Vec3::new(gen_non_zero(), gen_non_zero(), gen_non_zero()),
             rng.gen::<Quat>(),
             Vec3::new(
@@ -372,16 +376,30 @@ impl Distribution<TransformSrt> for Standard {
     }
 }
 
-impl From<TransformSrt> for Mat4 {
+impl From<TransformSRT> for Mat4 {
     #[inline]
-    fn from(srt: TransformSrt) -> Self {
-        Mat4::from_scale_rotation_translation(srt.scale, srt.rotation, srt.translation)
+    fn from(srt: TransformSRT) -> Self {
+        Self::from_scale_rotation_translation(srt.scale, srt.rotation, srt.translation)
     }
 }
 
-impl From<TransformRt> for Mat4 {
+impl From<TransformRT> for Mat4 {
     #[inline]
-    fn from(rt: TransformRt) -> Self {
-        Mat4::from_rotation_translation(rt.rotation, rt.translation)
+    fn from(rt: TransformRT) -> Self {
+        Self::from_rotation_translation(rt.rotation, rt.translation)
+    }
+}
+
+impl From<TransformSRT> for Affine3A {
+    #[inline]
+    fn from(srt: TransformSRT) -> Self {
+        Self::from_scale_rotation_translation(srt.scale, srt.rotation, srt.translation)
+    }
+}
+
+impl From<TransformRT> for Affine3A {
+    #[inline]
+    fn from(rt: TransformRT) -> Self {
+        Self::from_rotation_translation(rt.rotation, rt.translation)
     }
 }